entrega_final/requirements.txt
django-rest-framework
asgiref
crispy-bootstrap5
distlib
Django
django-crispy-forms
djangorestframework
env
filelock
mysql
mysqlclient
Pillow
pip
platformdirs
psycopg2-binary
pytz
setuptools
six
sqlparse
tzdata
virtualenv